# Hash table variants and optimizations