--- a/velox/common/base/SimdUtil-inl.h
+++ b/velox/common/base/SimdUtil-inl.h
@@ -1253,4 +1253,25 @@ xsimd::batch<T, A> reinterpretBatch(xsimd::batch<U, A> data, const A& arch) {
   return detail::ReinterpretBatch<T, U, A>::apply(data, arch);
 }
 
+inline bool memEqualUnsafe(const void* x, const void* y, int32_t size) {
+  constexpr int32_t kBatch = xsimd::batch<uint8_t>::size;
+
+  auto left = reinterpret_cast<const uint8_t*>(x);
+  auto right = reinterpret_cast<const uint8_t*>(y);
+  while (size > 0) {
+    auto bits = toBitMask(
+        xsimd::batch<uint8_t>::load_unaligned(left) ==
+        xsimd::batch<uint8_t>::load_unaligned(right));
+    if (bits == allSetBitMask<uint8_t, xsimd::default_arch>()) {
+      left += kBatch;
+      right += kBatch;
+      size -= kBatch;
+      continue;
+    }
+    auto leading = __builtin_ctz(~bits);
+    return leading >= size;
+  }
+  return true;
+}
+
 } // namespace facebook::velox::simd