--- a/velox/common/base/SimdUtil.h
+++ b/velox/common/base/SimdUtil.h
@@ -413,6 +413,9 @@ inline bool isDense(const T* values, int32_t size) {
 // Reinterpret batch of U into batch of T.
 template <typename T, typename U, typename A = xsimd::default_arch>
 xsimd::batch<T, A> reinterpretBatch(xsimd::batch<U, A>, const A& = {});
+// Compares memory at 'x' and 'y' and returns true if 'size' leading bytes are
+// equal. May address up to SIMD width -1 past end of either 'x' or 'y'.
+inline bool memEqualUnsafe(const void* x, const void* y, int32_t size);
 
 } // namespace facebook::velox::simd